backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-hcl2
redis
requests
pydantic
python-hcl2
networkx
python-multipart
PyYAML
